refactor use-cases and api gateways
