--- client_end_user/shared/src/commonMain/kotlin/data/gateway/remote/CartGateway.kt
@@ -1,10 +1,0 @@
-package data.gateway.remote
-
-import domain.entity.Cart
-import domain.gateway.ICartGateway
-
-class CartGateway : ICartGateway {
-    override suspend fun getAllCartMeals(): Cart {
-        TODO("Not yet implemented")
-    }
-}
