+++ client_end_user/shared/src/commonMain/kotlin/data/gateway/remote/OrderGateway.kt
@@ -1,5 +1,6 @@
 package data.gateway.remote
 
+import domain.entity.Cart
 import domain.entity.Order
 import domain.entity.Trip
 import domain.gateway.IOrderGateway
@@ -10,6 +11,10 @@ class OrderGateway : IOrderGateway {
     }
 
     override suspend fun getOrderHistoryGateway(): List<Order> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun getAllCartMeals(): Cart {
         TODO("Not yet implemented")
     }
 }
