+++ client_end_user/shared/src/commonMain/kotlin/data/gateway/remote/UserGateway.kt
@@ -3,6 +3,8 @@ package data.gateway.remote
 import data.remote.mapper.toEntity
 import data.remote.mapper.toSessionEntity
 import data.remote.mapper.toUser
+import data.remote.model.LocationDto
+import data.remote.model.RestaurantDto
 import data.remote.model.ServerResponse
 import data.remote.model.SessionDto
 import data.remote.model.UserDetailsDto
@@ -13,6 +15,7 @@ import domain.entity.UserCreation
 import domain.entity.UserDetails
 import domain.gateway.IUserGateway
 import domain.utils.AuthorizationException
+import domain.utils.GeneralException
 import io.ktor.client.HttpClient
 import io.ktor.client.request.forms.submitForm
 import io.ktor.client.request.get
@@ -89,4 +92,80 @@ class UserGateway(client: HttpClient) : BaseGateway(client), IUserGateway {
         // todo : implement this when the backend is ready
         return User(name = "Test", currency = "$", walletValue = 100.0)
     }
+
+    override suspend fun getFavoriteRestaurants(): List<domain.entity.Restaurant> {
+        return restaurants.map { it.toEntity() }
+    }
+
+    override suspend fun addRestaurantToFavorites(restaurantId: String): Boolean {
+        return tryToExecute<ServerResponse<Boolean>> {
+            submitForm(
+                url = ("/user/favorite"),
+                formParameters = Parameters.build {
+                    append("restaurantId", restaurantId)
+                }
+            ) {
+                method = HttpMethod.Post
+            }
+        }.value ?: throw GeneralException.NotFoundException
+    }
+
+    override suspend fun removeRestaurantFromFavorites(restaurantId: String): Boolean {
+        return tryToExecute<ServerResponse<Boolean>> {
+            submitForm(
+                url = ("/user/favorite"),
+                formParameters = Parameters.build {
+                    append("restaurantId", restaurantId)
+                }
+            ) {
+                method = HttpMethod.Delete
+            }
+        }.value ?: throw GeneralException.NotFoundException
+    }
+
+    // Should be removed when the backend is ready
+    private val restaurants = listOf(
+        RestaurantDto(
+            id = "64f372095fecc11e6d917656",
+            ownerId = "64f3663e5ddbc15bfd1efcfa",
+            ownerUsername = "hamada",
+            name = "Hamada Market",
+            rate = 4.5,
+            priceLevel = "$$",
+            phone = "1234567890",
+            description = "This is description",
+            closingTime = "07:30",
+            openingTime = "22:00",
+            location = LocationDto(22.0, 10.5),
+            address = "Main street, 123"
+        ),
+        RestaurantDto(
+            id = "64f373b35fecc11e6d917659",
+            ownerId = "64f3663e5ddbc15bfd1efcfa",
+            ownerUsername = "masala",
+            name = "Masala Restaurant",
+            rate = 3.5,
+            priceLevel = "$",
+            phone = "1234567890",
+            description = "This is description",
+            closingTime = "09:00",
+            openingTime = "20:00",
+            location = LocationDto(12.0, 10.5),
+            address = "New street, 23"
+        ),
+        RestaurantDto(
+            id = "64f373be5fecc11e6d91765a",
+            ownerId = "64f3663e5ddbc15bfd1efcfa",
+            ownerUsername = "the_chance",
+            name = "Chance Market",
+            rate = 5.0,
+            priceLevel = "$$$",
+            phone = "1234567890",
+            description = "This is description",
+            closingTime = "08:30",
+            openingTime = "17:00",
+            location = LocationDto(33.0, 22.5),
+            address = "Cairo street, 101"
+        )
+    )
 }
