--- client_end_user/shared/src/commonMain/kotlin/domain/gateway/ICartGateway.kt
@@ -1,7 +1,0 @@
-package domain.gateway
-
-import domain.entity.Cart
-
-interface ICartGateway {
-    suspend fun getAllCartMeals(): Cart
-}
