--- client_end_user/shared/src/commonMain/kotlin/domain/gateway/IFavoriteRestaurantGateway.kt
@@ -1,9 +1,0 @@
-package domain.gateway
-
-import domain.entity.Restaurant
-
-interface IFavoriteRestaurantGateway {
-    suspend fun getFavoriteRestaurants(): List<Restaurant>
-    suspend fun addRestaurantToFavorites(restaurantId: String): Boolean
-    suspend fun removeRestaurantFromFavorites(restaurantId: String): Boolean
-}
