--- client_end_user/shared/src/commonMain/kotlin/domain/gateway/IOffersGateway.kt
@@ -1,10 +1,0 @@
-package domain.gateway
-
-import domain.entity.Meal
-import domain.entity.Offer
-
-interface IOffersGateway {
-    fun getNewOffers(): List<Offer>
-    suspend fun getMostOrdersMeal(restaurantId: String): List<Meal>
-    suspend fun getSweets(restaurantId: String): List<Meal>
-}
