+++ client_end_user/shared/src/commonMain/kotlin/domain/gateway/IOrderGateway.kt
@@ -1,9 +1,11 @@
 package domain.gateway
 
+import domain.entity.Cart
 import domain.entity.Order
 import domain.entity.Trip
 
 interface IOrderGateway {
     suspend fun getTripHistory(): List<Trip>
     suspend fun getOrderHistoryGateway(): List<Order>
+    suspend fun getAllCartMeals(): Cart
 }
