+++ client_end_user/shared/src/commonMain/kotlin/domain/gateway/IRestaurantGateway.kt
@@ -3,6 +3,7 @@ package domain.gateway
 import domain.entity.Cuisine
 import domain.entity.InProgressWrapper
 import domain.entity.Meal
+import domain.entity.Offer
 import domain.entity.Restaurant
 
 interface IRestaurantGateway {
@@ -13,4 +14,8 @@ interface IRestaurantGateway {
     suspend fun getRestaurantDetails(restaurantId: String): Restaurant
 
     suspend fun getMealById(mealId: String): Meal
+
+    suspend fun getNewOffers(): List<Offer>
+    suspend fun getMostOrdersMeal(restaurantId: String): List<Meal>
+    suspend fun getSweets(restaurantId: String): List<Meal>
 }
