+++ client_end_user/shared/src/commonMain/kotlin/domain/gateway/IUserGateway.kt
@@ -1,5 +1,6 @@
 package domain.gateway
 
+import domain.entity.Restaurant
 import domain.entity.Session
 import domain.entity.User
 import domain.entity.UserCreation
@@ -15,4 +16,8 @@ interface IUserGateway {
     suspend fun getUserProfile(): UserDetails
 
     suspend fun getUserWallet(): User
+
+    suspend fun getFavoriteRestaurants(): List<Restaurant>
+    suspend fun addRestaurantToFavorites(restaurantId: String): Boolean
+    suspend fun removeRestaurantFromFavorites(restaurantId: String): Boolean
 }
