--- client_end_user/shared/src/commonMain/kotlin/domain/usecase/GetCuisinesUseCase.kt
@@ -1,14 +1,0 @@
-package domain.usecase
-
-import domain.entity.Cuisine
-import domain.gateway.IRestaurantGateway
-
-interface IGetCuisinesUseCase {
-    suspend fun getCuisines(): List<Cuisine>
-}
-
-class GetCuisinesUseCase(private val cuisineGateway: IRestaurantGateway) : IGetCuisinesUseCase {
-    override suspend fun getCuisines(): List<Cuisine> {
-        return cuisineGateway.getCuisines()
-    }
-}
