+++ client_end_user/shared/src/commonMain/kotlin/domain/usecase/IManageFavouriteUseCase.kt
@@ -1,7 +1,8 @@
 package domain.usecase
 
+import data.gateway.remote.UserGateway
 import domain.entity.Restaurant
-import domain.gateway.IFavoriteRestaurantGateway
+import domain.gateway.IUserGateway
 
 interface IManageFavouriteUseCase {
     suspend fun addRestaurantToFavorites(restaurantId: String): Boolean
@@ -10,17 +11,17 @@ interface IManageFavouriteUseCase {
 }
 
 class ManageFavouriteUseCase(
-    private val favoriteRestaurantGateway: IFavoriteRestaurantGateway
+    private val userGateway: IUserGateway
 ) : IManageFavouriteUseCase {
     override suspend fun addRestaurantToFavorites(restaurantId: String): Boolean {
-        return favoriteRestaurantGateway.addRestaurantToFavorites(restaurantId)
+        return userGateway.addRestaurantToFavorites(restaurantId)
     }
 
     override suspend fun removeRestaurantFromFavorites(restaurantId: String): Boolean {
-        return favoriteRestaurantGateway.removeRestaurantFromFavorites(restaurantId)
+        return userGateway.removeRestaurantFromFavorites(restaurantId)
     }
 
     override suspend fun getFavoriteRestaurants(): List<Restaurant> {
-        return favoriteRestaurantGateway.getFavoriteRestaurants()
+        return userGateway.getFavoriteRestaurants()
     }
 }
