+++ client_end_user/shared/src/commonMain/kotlin/domain/usecase/ManageCartUseCase.kt
@@ -1,20 +1,22 @@
 package domain.usecase
 
-import data.gateway.fake.FakeCartGateway
 import domain.entity.Cart
+import domain.gateway.IOrderGateway
 
 interface IManageCartUseCase {
     suspend fun getAllCartMeals(): Cart
     suspend fun checkIfThereIsOrderInCart(): Boolean
 }
 
-class ManageCartUseCase(private val remoteGateway: FakeCartGateway) : IManageCartUseCase {
+class ManageCartUseCase(
+    private val orderGateway: IOrderGateway
+) : IManageCartUseCase {
     override suspend fun getAllCartMeals(): Cart {
-        return remoteGateway.getAllCartMeals()
+        return orderGateway.getAllCartMeals()
     }
 
     override suspend fun checkIfThereIsOrderInCart(): Boolean {
-       val cart = remoteGateway.getAllCartMeals()
+        val cart = orderGateway.getAllCartMeals()
         return cart.meals.isEmpty()
     }
 }
