+++ client_end_user/shared/src/commonMain/kotlin/domain/usecase/ManageUserUseCase.kt
@@ -6,16 +6,13 @@ import data.local.mapper.toPreferredRide
 import domain.entity.PreferredRide
 import domain.entity.User
 import domain.entity.UserDetails
+import domain.gateway.IUserGateway
 import domain.gateway.local.ILocalConfigurationGateway
 import domain.utils.AuthorizationException
 import kotlinx.coroutines.flow.Flow
 
 interface IManageUserUseCase {
     suspend fun getUserWallet(): User
-    suspend fun savePriceLevel(priceLevel: String)
-    suspend fun getPriceLevel(): String
-    suspend fun savePreferredRide(preferredRide: PreferredRide)
-    suspend fun getPreferredRide(): PreferredRide
     suspend fun saveIsFirstTimeUseApp(isFirstTimeUseApp: Boolean)
     suspend fun getIsFirstTimeUseApp(): Boolean
     suspend fun getUserLanguageCode(): Flow<String>
@@ -24,7 +21,7 @@ interface IManageUserUseCase {
 
 class ManageUserUseCase(
     private val localGateway: ILocalConfigurationGateway,
-    private val userGateway: UserGateway
+    private val userGateway: IUserGateway
 ) : IManageUserUseCase {
 
     override suspend fun getUserWallet(): User {
@@ -33,22 +30,6 @@ class ManageUserUseCase(
         } else {
             throw AuthorizationException.UnAuthorizedException
         }
-    }
-
-    override suspend fun savePriceLevel(priceLevel: String) {
-        localGateway.savePriceLevel(priceLevel)
-    }
-
-    override suspend fun getPriceLevel(): String {
-        return localGateway.getPriceLevel()
-    }
-
-    override suspend fun savePreferredRide(preferredRide: PreferredRide) {
-        localGateway.savePreferredRideQuality(preferredRide.toFormattedString())
-    }
-
-    override suspend fun getPreferredRide(): PreferredRide {
-        return localGateway.getPreferredRideQuality().toPreferredRide()
     }
 
     override suspend fun saveIsFirstTimeUseApp(isFirstTimeUseApp: Boolean) {
