+++ client_end_user/shared/src/commonMain/kotlin/domain/usecase/mangeResturantUseCase.kt
@@ -1,5 +1,6 @@
 package domain.usecase
 
+import domain.entity.Cuisine
 import domain.entity.Meal
 import domain.entity.Restaurant
 import domain.gateway.IRestaurantGateway
@@ -7,16 +8,22 @@ import domain.gateway.IRestaurantGateway
 interface IMangeRestaurantUseCase {
     suspend fun getRestaurantDetails(restaurantId: String): Restaurant
     suspend fun getMealById(mealId: String): Meal
+    suspend fun getCuisines(): List<Cuisine>
+
 }
 
 class MangeRestaurantUseCase(
-    private val restaurantRemoteGateway: IRestaurantGateway,
+    private val restaurantGateway: IRestaurantGateway,
 ) : IMangeRestaurantUseCase {
     override suspend fun getRestaurantDetails(restaurantId: String): Restaurant {
-        return restaurantRemoteGateway.getRestaurantDetails(restaurantId)
+        return restaurantGateway.getRestaurantDetails(restaurantId)
     }
 
     override suspend fun getMealById(mealId: String): Meal {
-        return restaurantRemoteGateway.getMealById(mealId)
+        return restaurantGateway.getMealById(mealId)
+    }
+
+    override suspend fun getCuisines(): List<Cuisine> {
+        return restaurantGateway.getCuisines()
     }
 }
