+++ client_end_user/shared/src/commonMain/kotlin/presentation/home/HomeScreenModel.kt
@@ -4,12 +4,12 @@ import cafe.adriel.voyager.core.model.coroutineScope
 import domain.entity.InProgressWrapper
 import domain.entity.Restaurant
 import domain.entity.User
-import domain.usecase.IGetCuisinesUseCase
 import domain.usecase.IInProgressTrackerUseCase
 import domain.usecase.IManageCartUseCase
 import domain.usecase.IManageFavouriteUseCase
 import domain.usecase.IManageOffersUseCase
 import domain.usecase.IManageUserUseCase
+import domain.usecase.IMangeRestaurantUseCase
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import presentation.base.BaseScreenModel
@@ -18,7 +18,7 @@ import presentation.cuisines.CuisineUiState
 import presentation.cuisines.toCuisineUiState
 
 class HomeScreenModel(
-    private val cuisineUseCase: IGetCuisinesUseCase,
+    private val manageRestaurant: IMangeRestaurantUseCase,
     private val offers: IManageOffersUseCase,
     private val inProgressTrackerUseCase: IInProgressTrackerUseCase,
     private val manageUserUseCase: IManageUserUseCase,
@@ -127,7 +127,7 @@ class HomeScreenModel(
 
     private fun getRecommendedCuisines() {
         tryToExecute(
-            { cuisineUseCase.getCuisines().toCuisineUiState() },
+            { manageRestaurant.getCuisines().toCuisineUiState() },
             ::onGetCuisinesSuccess,
             ::onGetCuisinesError
         )
@@ -174,5 +174,4 @@ class HomeScreenModel(
     private fun onGetNewOffersError(error: ErrorState) {
         println("error is $error")
     }
-
 }
