+++ client_end_user/shared/src/commonMain/kotlin/presentation/preferredMeal/PreferredScreenModel.kt
@@ -2,15 +2,15 @@ package presentation.preferredMeal
 
 import cafe.adriel.voyager.core.model.coroutineScope
 import domain.usecase.IManageUserUseCase
-import domain.usecase.ManageUserUseCase
+import domain.usecase.IMangeUserPreferenceUseCase
 import kotlinx.coroutines.CoroutineScope
 import presentation.base.BaseScreenModel
 import presentation.base.ErrorState
 
 class PreferredScreenModel(
-    private val userPreference: ManageUserUseCase,
-    private val manageUser: IManageUserUseCase
-) :
+    private val manageUser: IManageUserUseCase,
+    private val userPreferences: IMangeUserPreferenceUseCase,
+    ) :
     BaseScreenModel<PreferredScreenUiState, PreferredScreenUiEffect>(PreferredScreenUiState()),
     PreferredScreenInteractionListener {
 
@@ -23,7 +23,7 @@ class PreferredScreenModel(
 
     private fun savePriceLevel(priceLevel: String) {
         tryToExecute(
-            { userPreference.savePriceLevel(priceLevel) },
+            { userPreferences.savePriceLevel(priceLevel) },
             ::onSavePriceLevelSuccess,
             ::onSavePriceLevelError
         )
